# tests/baseline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOAR Lab Management API",
    "description": "API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/analytics/kpis": {
      "get": {
        "summary": "Get Kpis",
        "description": "Get KPI metrics.\n\nArgs:\n    log_file_path: Optional path to a log file used for KPI\n    computation; when omitted the analytics service default is used.\n    analytics: The analytics service dependency.\n\nReturns:\n    dict: The comprehensive KPI metrics computed by the analytics\n    service.\n\nRaises:\n    HTTPException: 500 when KPI retrieval fails.",
        "operationId": "get_kpis_analytics_kpis_get",
        "parameters": [
          {
            "name": "log_file_path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Log File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Kpis Analytics Kpis Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/kpis/aggregated": {
      "get": {
        "summary": "Get Aggregated Kpis",
        "description": "Get aggregated KPI metrics from Elasticsearch soar-metrics index.\n\nArgs:\n    hours: The lookback window in hours for aggregating metrics.\n\nReturns:\n    dict: Aggregated KPI metrics including ``period_hours``,\n    ``timestamp``, ``total_alerts``, ``mttr_statistics``,\n    ``by_alert_type`` and ``services``.\n\nRaises:\n    HTTPException: 500 when retrieving or computing aggregated KPIs\n    fails.",
        "operationId": "get_aggregated_kpis_analytics_kpis_aggregated_get",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 24,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Aggregated Kpis Analytics Kpis Aggregated Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/node-timings": {
      "get": {
        "summary": "Get Node Timings",
        "description": "Get per-node timing breakdown for workflow executions.\n\nIf execution_id is provided, fetches timings for that specific execution.\nOtherwise, returns aggregated node timings from the last `hours` hours.\n\nArgs:\n    execution_id: Optional workflow execution identifier to fetch\n    timings for a single execution.\n    alert_id: Optional alert identifier used alongside execution_id.\n    hours: The lookback window in hours for aggregated timings.\n\nReturns:\n    dict: The per-node timing data for the requested execution or the\n    aggregated timings over the lookback window.\n\nRaises:\n    HTTPException: 404 when the requested execution is not found, or\n    500 when retrieving node timings fails.",
        "operationId": "get_node_timings_analytics_node_timings_get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Execution Id"
            }
          },
          {
            "name": "alert_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Alert Id"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Node Timings Analytics Node Timings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contain": {
      "post": {
        "summary": "Contain Endpoint",
        "description": "Simulated containment endpoint for the SOAR workflow.\n\nRecords a containment action without executing real firewall or endpoint\ncommands. Used by the Shuffle workflow when the decision node classifies\nthe alert as malicious.\n\nArgs:\n    request: The ContainmentRequest describing the alert, case, host\n    and containment mode.\n\nReturns:\n    ContainmentResponse: The simulated containment result including\n    the actions that would have been performed and a timestamp.\n\nRaises:\n    HTTPException: 500 when an unexpected error occurs during\n    containment handling.",
        "operationId": "contain_endpoint_api_v1_contain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cache/ioc": {
      "post": {
        "summary": "Cache Ioc Endpoint",
        "description": "Cache an IoC key/value in Redis on behalf of the Shuffle workflow.\n\nRedis speaks the RESP protocol, not HTTP, so Shuffle's generic \"http\"\napp node cannot POST JSON directly to port 6379 (it gets a\nconnection reset). This endpoint proxies the SET through the real\nredis-py client configured on the API service.\n\nArgs:\n    request: The IocCacheRequest carrying the key, value and optional\n    TTL to cache.\n\nReturns:\n    IocCacheResponse: Whether the cache operation succeeded along with\n    the key and a status message.",
        "operationId": "cache_ioc_endpoint_api_v1_cache_ioc_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IocCacheRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IocCacheResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContainmentRequest": {
        "properties": {
          "alert_id": {
            "type": "string",
            "title": "Alert Id",
            "description": "Alert identifier"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id",
            "description": "TheHive case identifier"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname",
            "description": "Target hostname to contain"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Containment mode",
            "default": "simulated"
          }
        },
        "type": "object",
        "required": [
          "alert_id",
          "case_id",
          "hostname"
        ],
        "title": "ContainmentRequest",
        "description": "Request model for the simulated containment endpoint."
      },
      "ContainmentResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Containment status"
          },
          "alert_id": {
            "type": "string",
            "title": "Alert Id",
            "description": "Alert identifier"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id",
            "description": "TheHive case identifier"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname",
            "description": "Target hostname"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions",
            "description": "Actions performed"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO format timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "alert_id",
          "case_id",
          "hostname",
          "timestamp"
        ],
        "title": "ContainmentResponse",
        "description": "Response model for the simulated containment endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IocCacheRequest": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Cache key"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "Cache value"
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Seconds",
            "description": "Optional TTL in seconds"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "IocCacheRequest",
        "description": "Request model for the IoC cache endpoint."
      },
      "IocCacheResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the cache operation succeeded"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Cache key"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "success",
          "key"
        ],
        "title": "IocCacheResponse",
        "description": "Response model for the IoC cache endpoint."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}